add more propreties for medicament entity

+++ PharmaHub.API/Contracts/Medicament/CreateMedicamentRequest.cs
@@ -9,11 +9,11 @@ public class CreateMedicamentRequest
     public string Form { get; set; }
     public double PPV { get; set; }
     public double PPH { get; set; }
-    public double TVA { get; set; }
+    public float TVA { get; set; }
     public float Discount { get; set; }
     public double PBR { get; set; }
     public string Type { get; set; }
-    public int Marge { get; set; }
+    public double Marge { get; set; }
     public string Codebar { get; set; }
     public string Family { get; set; }
     [JsonConverter(typeof(JsonStringEnumConverter<UsedBy>))]
